backend/requirements.txt
# Django Core
Django==5.0.1
psycopg2-binary==2.9.9

# Django REST Framework
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1

# CORS Headers
django-cors-headers==4.3.1

# Environment variables
python-decouple==3.8

# Production server
gunicorn==21.2.0

# Development tools
django-debug-toolbar==4.2.0

# Code quality
black==24.1.1
flake8==7.0.0
isort==5.13.2
